break rbldap.search in to its own file
